refactored menu functions into their related menu files
